Change the setter of setPrice of Bag : int > float

+++ src/Entity/Bag.php
@@ -77,7 +77,7 @@ class Bag
         return $this->price;
     }
 
-    public function setPrice(int $price) : self
+    public function setPrice(float $price) : self
     {
         $this->price = $price;
         return $this;
